feat: pwa hook
