+++ src/shared/hooks/use-pwa.ts
@@ -6,8 +6,19 @@ declare global {
   }
 }
 
+interface BeforeInstallPromptEvent extends Event {
+  readonly platforms: Array<string>
+  readonly userChoice: Promise<{
+    outcome: 'accepted' | 'dismissed'
+    platform: string
+  }>
+  prompt(): Promise<void>
+}
+
 export const usePWA = () => {
   const [isPWA, setIsPWA] = useState(false)
+  const [installPrompt, setInstallPrompt] = useState<BeforeInstallPromptEvent | null>(null)
+  const [installable, setInstallable] = useState(false)
 
   useEffect(() => {
     const checkPWA = () => {
@@ -27,5 +38,43 @@ export const usePWA = () => {
     return () => mediaQuery.removeEventListener('change', handler)
   }, [])
 
-  return { isPWA }
+  useEffect(() => {
+    const handleBeforeInstallPrompt = (e: BeforeInstallPromptEvent) => {
+      e.preventDefault()
+      setInstallPrompt(e)
+      setInstallable(true)
+    }
+
+    window.addEventListener('beforeinstallprompt', handleBeforeInstallPrompt as EventListener)
+
+    return () => {
+      window.removeEventListener('beforeinstallprompt', handleBeforeInstallPrompt as EventListener)
+    }
+  }, [])
+
+  const installPWA = async () => {
+    if (installPrompt) {
+      try {
+        await installPrompt.prompt()
+        const result = await installPrompt.userChoice
+
+        if (result.outcome === 'accepted') {
+          console.log('앱 설치 승인')
+        } else {
+          console.log('앱 설치 거부')
+        }
+
+        setInstallable(false)
+      } catch (error) {
+        // 추가 폴백 메커니즘
+        console.error('설치 중 오류 발생', error)
+
+        alert(
+          '앱 설치 과정에서 오류가 발생했습니다. 다시 시도해주세요. 문제가 반복될 경우 브라우저의 ⋮ (더보기) 버튼 혹은 공유 버튼을 클릭 후 "홈 화면에 추가" 옵션을 클릭하시면 앱 설치가 가능합니다.',
+        )
+      }
+    }
+  }
+
+  return { isPWA, installable, installPWA }
 }
